requirements.txt
myo_python==1.0.4
tensorflow==2.4.1
tensorflow_addons==0.12.1
Keras==2.4.3
firebase_admin==4.5.3
matplotlib==3.4.1
pandas==1.2.4
keyboard==0.13.5
seaborn==0.11.1
numpy==1.19.5
pynput==1.7.3
PyQt6==6.1.0
